Update Bootstrap.php
Under  case 'Zend_Cache_Backend_File':  (Approx line 265)

File:
	Application/Bootstrap/Bootstrap.php

Line 265
	default:
	$dir = $wordPress->getCustomOption(
		'cacheFile', 
		tempnam(sys_get_temp_dir(), 'vulnero')
        );
	if (is_file($dir)) {
		unlink($dir);
	}
                
	if (!is_dir($dir)) {
		mkdir($dir, 0775);
	}


The problem is that the 2nd parameter to getCustomOption is
the default value.  It's only used if the "cacheFile" option
doesn't exist.

If the "cacheFile" option exists, then, the value returned 
from tempnam() is never used for anything.

The tempnam() function generates a file with a unique filename:
	"Creates a file with a unique filename, with access permission set to 0600, in the specified directory."

	//see documentation:
	http://php.net/manual/en/function.tempnam.php

If you follow the code above, you'll see that the 
file that tempNam generates is never deleted.  It
just gives you a new unique filename everytime 
the code is executed.

On our webserver, it kept generating new files, until
it went over 65k files.  Eventually, I start getting this error:
	PHP Notice:  tempnam(): file created in the system's temporary directory in...

	//see documentation here: http://php.net/manual/en/function.tempnam.php
	Note: If PHP cannot create a file in the specified dir parameter, it falls back on the system default. On NTFS this also happens if the specified dir contains more than 65534 files.